% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimPercen.R
\name{GetSimPercen}
\alias{GetSimPercen}
\title{This function is used to calculate the similarities between
two protein sequences with a percentage output.}
\arguments{
\item{protein1}{A protein Uniprot}

\item{protein2}{A protein Uniprot}
}
\value{
Return a percentage number tells the
similarities between selected protein sequences
}
\description{
This function is used to calculate the similarities between
two protein sequences with a percentage output.
}
\examples{
GetSimilarPercentage("Q9UHB7", "Q9UKV5")

}
