% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUA.R
\name{runUA}
\alias{runUA}
\title{Launch the shiny app for package UbiquitinAnalysis}
\usage{
runUA()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for the package
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runUA()
}
}
