\name{GetSimilarPercentage}
\alias{GetSimilarPercentage}
\title{
Get Selected Protein Sequences Similarities.
}
\description{
Input two selected proteins and get their sequence similarities in percentage.
}
\usage{
GetSimilarPercentage(protein1, protein2);
}
\value{
Returns a percentage number to tell these two sequences similarities.
}

\arguments{
  \item{protein1}{}
  \item{protein2}{}
}

\examples{
GetSimilarPercentage("Q9UKV5", "Q9UHB7");
}

